Initial integration of https://gitlab.com/visr/Zarr.jl
Moved the code in, but did not yet hook it up completely.
Added some tests and a minimal README.

Next planned step is to include a Metadata instance as a ZArray field, 
and integrate it further.

Note that this sacrifices FixedLengthUTF8Strings for now. It also 
removes the last type parameter of ZArray, which was the same as T 
except for FixedLengthUTF8Strings.

ref #1