Refactor compressors into multiple files
Refactors each compressor into a single file and the abstract API + NoCompressor to `Compressors/Compressors.jl`.  Also adds API docs for the Compressor API in case people want to do that.

Future work may also explore making this compliant with Interfaces.jl so that we have a way to test that all compressors are compliant.